Start at trying to export changes back to wkit.
should parse through the json and change the position data on the nodes based on the items in the blender tree that have the tags from the importer. can only move stuff so far, and doesnt work on instances. Havent actually got it to work in game yet. Just putting the stake in the sand as the starting point.